--- functions/firebase.json
@@ -1,24 +1,0 @@
-{
-  "database": {
-    "rules": "security.rules.json"
-  },
-  "hosting": {
-    "public": "./public",
-    "headers": [
-      {
-        "source": "**/*.@(js|html)",
-        "headers": [
-          {
-            "key": "Cache-Control",
-            "value": "max-age=0"
-          }
-        ]
-      }
-    ]
-  },
-  "functions": {
-    "predeploy": [
-      "npm --prefix \"$RESOURCE_DIR\" run lint"
-    ]
-  }
-}
