--- functions/functions/package.json
@@ -1,26 +1,0 @@
-{
-  "name": "functions",
-  "description": "Cloud Functions for Firebase",
-  "scripts": {
-    "lint": "eslint .",
-    "serve": "firebase serve --only functions",
-    "shell": "firebase experimental:functions:shell",
-    "start": "npm run shell",
-    "deploy": "firebase deploy --only functions",
-    "logs": "firebase functions:log"
-  },
-  "dependencies": {
-    "firebase-admin": "~8.5.0",
-    "firebase-functions": "^3.2.0",
-    "bad-words": "^1.3.1",
-    "capitalize-sentence": "^0.1.2"
-  },
-  "devDependencies": {
-    "eslint": "^4.12.0",
-    "eslint-plugin-promise": "^3.6.0"
-  },
-  "engines": {
-    "node": "8"
-  },
-  "private": true
-}
