--- functions/public/index.html
@@ -1,137 +1,0 @@
-<!doctype html>
-<!--
-  Copyright 2016 Google Inc. All rights reserved.
-  Licensed under the Apache License, Version 2.0 (the "License");
-  you may not use this file except in compliance with the License.
-  You may obtain a copy of the License at
-      https://www.apache.org/licenses/LICENSE-2.0
-  Unless required by applicable law or agreed to in writing, software
-  distributed under the License is distributed on an "AS IS" BASIS,
-  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  See the License for the specific language governing permissions and
-  limitations under the License
--->
-<html lang="en">
-<head>
-  <meta charset="utf-8">
-  <meta http-equiv="X-UA-Compatible" content="IE=edge">
-  <meta name="description" content="Demonstrates the use of Google Cloud Database with a Firebase DB">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>HTTPS Callable functions Web Quickstart</title>
-
-  <!-- Disable tap highlight on IE -->
-  <meta name="msapplication-tap-highlight" content="no">
-
-  <!-- Add to homescreen for Chrome on Android -->
-  <meta name="mobile-web-app-capable" content="yes">
-  <meta name="application-name" content="Firebase Database Quickstart">
-  <meta name="theme-color" content="#303F9F">
-
-  <!-- Add to homescreen for Safari on iOS -->
-  <meta name="apple-mobile-web-app-capable" content="yes">
-  <meta name="apple-mobile-web-app-status-bar-style" content="black-translucent">
-  <meta name="apple-mobile-web-app-title" content="Firebase Database Quickstart">
-  <meta name="apple-mobile-web-app-status-bar-style" content="#303F9F">
-
-  <!-- Tile icon for Win8 -->
-  <meta name="msapplication-TileColor" content="#3372DF">
-  <meta name="msapplication-navbutton-color" content="#303F9F">
-
-  <!-- Material Design Lite -->
-  <link rel="stylesheet" href="https://fonts.googleapis.com/icon?family=Material+Icons">
-  <link rel="stylesheet" href="https://code.getmdl.io/1.1.3/material.blue_grey-orange.min.css">
-  <script src="https://code.getmdl.io/1.1.3/material.min.js"></script>
-
-  <link rel="stylesheet" href="/styles/main.css">
-
-</head>
-<body>
-<div class="demo-layout mdl-layout mdl-js-layout mdl-layout--fixed-header">
-
-
-  <!-- Splash screen -->
-  <section id="page-splash">
-    <h3 class="logo">HTTPS Callable functions Web Quickstart</h3>
-    <div>
-      <button id="sign-in-button" class="mdl-button--raised mdl-button mdl-js-button mdl-js-ripple-effect"><i class="material-icons">account_circle</i> Sign in with Google</button>
-    </div>
-  </section>
-
-  <!-- Header section containing logo and menu -->
-  <header class="header mdl-layout__header mdl-color-text--white mdl-color--light-blue-700">
-
-    <div class="mdl-layout__header-row titlebar">
-      <h3 class="logo">HTTPS Callable functions Web Quickstart</h3>
-      <button id="sign-out-button" class="mdl-button--raised mdl-button mdl-js-button mdl-js-ripple-effect"><i class="material-icons">account_circle</i> Sign out</button>
-    </div>
-  </header>
-
-  <main class="mdl-layout__content mdl-color--grey-100">
-
-    <!-- Shows actions that uses Callable Cloud Functions -->
-    <section class="mdl-grid content">
-      <div class="mdl-cell mdl-cell--12-col mdl-grid">
-
-        <!-- Action that adds 2 numbers -->
-        <div class="mdl-card mdl-shadow--2dp mdl-cell mdl-cell--12-col mdl-cell--8-col-tablet
-                  mdl-cell--8-col-desktop">
-          <div class="mdl-card__title mdl-color--light-blue-600 mdl-color-text--white">
-            <h2 class="mdl-card__title-text">Add two numbers via a server-side function</h2>
-          </div>
-          <div class="mdl-card__supporting-text mdl-color-text--grey-600">
-            <p>Add two numbers using a server side function. Enter the two numbers:</p>
-            <form>
-              <div class="mdl-textfield mdl-js-textfield mdl-textfield--floating-label number-input">
-                <input class="mdl-textfield__input" pattern="-?[0-9]*(\.[0-9]+)?" id="first-number">
-                <label class="mdl-textfield__label" for="first-number">Number...</label>
-                <span class="mdl-textfield__error">Input is not a number!</span>
-              </div>
-              <div class="mdl-textfield mdl-js-textfield mdl-textfield--floating-label number-input">
-                <input class="mdl-textfield__input" pattern="-?[0-9]*(\.[0-9]+)?" id="second-number">
-                <label class="mdl-textfield__label" for="second-number">Number...</label>
-                <span class="mdl-textfield__error">Input is not a number!</span>
-              </div>
-              <button id="send-addition-button" class="mdl-button--raised mdl-button mdl-js-button mdl-js-ripple-effect">Add the two numbers</button>
-            </form>
-          </div>
-        </div>
-
-        <!-- Action that writes a message to the Realtime Database -->
-        <div class="mdl-card mdl-shadow--2dp mdl-cell mdl-cell--12-col mdl-cell--8-col-tablet
-                  mdl-cell--8-col-desktop">
-          <div class="mdl-card__title mdl-color--light-blue-600 mdl-color-text--white">
-            <h2 class="mdl-card__title-text">Write a message sanitized by a server-side function</h2>
-          </div>
-          <div class="mdl-card__supporting-text mdl-color-text--grey-600">
-            <p>Add a message to the database. The message will be checked for <a href="https://github.com/web-mech/badwords-list">bad words</a> or SHOUTING!! and sanitized:</p>
-            <form>
-              <div class="mdl-textfield mdl-js-textfield mdl-textfield--floating-label">
-                <input class="mdl-textfield__input" id="message-text">
-                <label class="mdl-textfield__label" for="message-text">Message text...</label>
-              </div>
-              <button id="add-message-button" class="mdl-button--raised mdl-button mdl-js-button mdl-js-ripple-effect">Add Message</button>
-            </form>
-          </div>
-
-          <div class="mdl-card__title mdl-color--light-blue-300 mdl-color-text--white">
-            <h2 class="mdl-card__title-text">Messages written</h2>
-          </div>
-          <div id="message-container"/>
-        </div>
-      </div>
-    </section>
-  </main>
-</div>
-
-<!-- Import and configure the Firebase SDK -->
-<!-- These scripts are made available when the app is served or deployed on Firebase Hosting -->
-<!-- If you do not serve/host your project using Firebase Hosting see https://firebase.google.com/docs/web/setup -->
-<script src="/__/firebase/6.5.0/firebase-app.js"></script>
-<script src="/__/firebase/6.5.0/firebase-auth.js"></script>
-<script src="/__/firebase/6.5.0/firebase-database.js"></script>
-<script src="/__/firebase/6.5.0/firebase-functions.js"></script>
-<script src="/__/firebase/init.js"></script>
-
-<script src="/scripts/main.js"></script>
-</body>
-</html>
