--- functions/public/styles/main.css
@@ -1,205 +1,0 @@
-/**
- * Copyright 2015 Google Inc. All Rights Reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-html, body {
-    font-family: 'Roboto', 'Helvetica', sans-serif;
-}
-a {
-    text-decoration: underline;
-    color: #039BE5;
-}
-.mdl-menu a {
-    text-decoration: none;
-}
-.content {
-    max-width: 1024px;
-}
-.mdl-card {
-    display: block;
-}
-.number-input {
-    max-width: 100px;
-    margin-right: 50px;
-}
-h3 {
-    background: url('/images/firebase-logo.png') no-repeat;
-    background-size: 40px;
-    padding-left: 50px;
-    background-position-y: -2px;
-}
-.note {
-    font-style: italic;
-    margin-top: 25px;
-}
-#disabled-message {
-    color: red;
-    margin-left: 20px;
-}
-.mdl-textfield {
-    width: 100%;
-}
-/* Fixes an MDL bug where the header does not scroll on small devices */
-.mdl-layout__container .mdl-layout--fixed-header .mdl-layout__content {
-    overflow-y: visible;
-    overflow-x: visible;
-    overflow: visible;
-}
-/* Overrides MDL colors */
-.mdl-layout.is-upgraded .mdl-layout__tab.is-active::after,
-.header .mdl-textfield__label:after {
-    background-color: #FFCA28;
-}
-.mdl-snackbar__action {
-    color: #FFCA28;
-}
-.mdl-textfield__label:after {
-    background-color: #0288D1;
-}
-.mdl-textfield--floating-label.is-focused .mdl-textfield__label {
-    color: #0288D1;
-}
-
-/* Header */
-
-.logo {
-    font-family: 'Amaranth', sans-serif;
-}
-.logo .material-icons {
-    top: 4px;
-    font-size: 32px;
-    margin-right: -2px;
-    position: relative;
-}
-.header .mdl-layout__header-row {
-    max-width: 1024px;
-    width: 100%;
-    height: auto;
-    padding: 10px 20px;
-    margin: auto;
-    position: relative;
-}
-@media screen and (max-width: 840px) {
-    .header .mdl-layout__header-row.titlebar {
-        height: 56px;
-    }
-    .titlebar h3 {
-        font-size: 24px;
-    }
-    .titlebar h3 .material-icons {
-        top: 2.5px;
-        font-size: 23px;
-    }
-}
-@media screen and (max-width: 479px) {
-    .mdl-grid {
-        padding: 0;
-    }
-    .mdl-card {
-        font-size: 14px;
-    }
-}
-#sign-in-button {
-    margin-left: 10px;
-    color: inherit;
-}
-#sign-out-button {
-    margin: 0 10px 10px 10px;
-    color: inherit;
-    right: 0;
-    position: absolute;
-    background-color: #328EC6;
-}
-.mdl-button .material-icons {
-    margin-top: -2px;
-}
-.mdl-button.mdl-button--icon .material-icons,
-.mdl-button.mdl-button--fab .material-icons {
-    margin-top: 0;
-}
-
-/* Navigation Bar */
-
-.header .mdl-tab {
-    padding-left: 23px;
-    box-sizing: border-box;
-    width: 100%;
-    height: 100%;
-    overflow: visible;
-    position: relative;
-    margin: auto;
-    max-width: 1024px;
-}
-@media screen and (max-width: 609px) {
-    .header .mdl-tab {
-        width: auto;
-    }
-}
-.header .tab {
-    padding: 0;
-    max-width: none;
-}
-#add {
-    position: absolute;
-    right: 20px;
-    top: 18px;
-    z-index: 998;
-}
-@media screen and (max-width: 840px) {
-    #add {
-        position: fixed;
-        bottom: 10px;
-        right: 10px;
-        z-index: 998;
-        top: auto;
-    }
-}
-
-
-/* Splash Page */
-
-#page-splash {
-    position: fixed;
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 100%;
-    background-color: #0288D1;
-    background: radial-gradient(circle, #039BE5, #01579b);
-    z-index: 10000;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    color: white;
-    flex-direction: column;
-}
-
-/* Messages */
-
-.message-element {
-    display: flex;
-    flex-direction: row;
-    padding: 10px
-}
-.message-element:nth-child(even) {
-    background-color: #eee;
-}
-.author-container:after {
-    content: ':';
-}
-.author-container {
-    margin-right: 5px;
-    font-weight: bold;
-}
