--- functions/security.rules.json
@@ -1,8 +1,0 @@
-{
-  "rules": {
-    "messages": {
-      ".read": true,
-      ".write": false
-    }
-  }
-}
